Benchmark programs for the lazy decoder added
